Format and fine tune the ProgressTracker
